restructure to react-redux-toolkit style
