+++ package.json
@@ -18,6 +18,7 @@
     "bootstrap": "^5.1.3",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
+    "prettier": "^2.5.1",
     "react": "^17.0.2",
     "react-bootstrap": "^2.0.3",
     "react-dom": "^17.0.2",
