--- src/App.test.tsx
@@ -1,8 +1,0 @@
-// import React from "react";
-import { render } from "@testing-library/react";
-import App from "./App";
-
-test("renders learn react link", () => {
-  const { container } = render(<App />);
-  expect(container.childElementCount).toBeGreaterThan(0);
-});
