+++ src/app/App.test.tsx
@@ -1,0 +1,7 @@
+import { render } from "@testing-library/react";
+import App from "./App";
+
+test("renders learn react link", () => {
+  const { container } = render(<App />);
+  expect(container.childElementCount).toBeGreaterThan(0);
+});
