+++ src/features/chat/components/message-list/message-list.component.tsx
@@ -1,0 +1,31 @@
+import { useEffect, useRef } from "react";
+import { Stack } from "react-bootstrap";
+import { Message } from "../../../../model/message";
+import { User } from "../../../../model/user";
+import { MessageComponent } from "../message/message.component";
+
+interface MessageListProps {
+  messages: Message[];
+  currentUser: User;
+}
+
+export const MessageListComponent: React.FC<MessageListProps> = (props) => {
+  const ref = useRef<HTMLDivElement>(null);
+  useEffect(() => {
+    if (ref.current) {
+      const scrollTo = ref.current.scrollHeight - ref.current.clientHeight;
+      ref.current.scrollTop = scrollTo;
+    }
+  }, [props.messages]);
+  return (
+    <Stack className="justify-content-end overflow-scroll m-2" ref={ref}>
+      {props.messages.map((message) => (
+        <MessageComponent
+          key={message.id}
+          message={message}
+          isCurrentUserAuthor={message.user === props.currentUser}
+        ></MessageComponent>
+      ))}
+    </Stack>
+  );
+};
