+++ src/features/chat/components/new-message/new-message.css
