+++ src/features/chat/new-message.ts
@@ -1,0 +1,6 @@
+import { Message } from "../../model/message";
+
+export interface NewMessage {
+  chatId: string;
+  message: Message;
+}
